+++ Gppbox/engine/Utils.cpp
@@ -12,7 +12,7 @@ bool Utils::isFullBody(Entity* entity, int gridx, int gridy)
 
 float Utils::toLength(const sf::Vector2i& source)
 {
-    return sqrt((source.x * source.x) + (source.y * source.y));
+    return (float)sqrt((source.x * source.x) + (source.y * source.y));
 }
 
 float Utils::toLength(const sf::Vector2f& source)
@@ -22,21 +22,21 @@ float Utils::toLength(const sf::Vector2f& source)
 
 float Utils::toAngle(const sf::Vector2i& source)
 {
-    float angle = atan2(source.y, source.x);
-    if (angle < 0.0f) angle += (M_PI + M_PI);
-    return angle * RadToDeg;
+    float angle = (float)atan2(source.y, source.x);
+    if (angle < 0.0f) angle += float(M_PI + M_PI);
+    return float(angle * RadToDeg);
 }
 
 float Utils::toAngle(const sf::Vector2f& source)
 {
     float angle = atan2(source.y, source.x);
-    if (angle < 0.0f) angle += (M_PI + M_PI);
-    return angle * RadToDeg;
+    if (angle < 0.0f) angle += float(M_PI + M_PI);
+    return float(angle * RadToDeg);
 }
 
 sf::Vector2f Utils::fromAngle(float angle)
 {
-    float rad = angle * DegToRad;
+    float rad = float(angle * DegToRad);
     return normalize(sf::Vector2f {
         cos(rad),
         sin(rad)
@@ -50,7 +50,7 @@ sf::Vector2f Utils::normalize(const sf::Vector2i& source)
 {
     float length = toLength(source);
     if (length != 0) return sf::Vector2f(source.x / length, source.y / length);
-    else return sf::Vector2f(source.x, source.y);
+    else return sf::Vector2f((float)source.x, (float)source.y);
 }
 
 sf::Vector2f Utils::normalize(const sf::Vector2f& source)
